[HW07] Refactor according to the linter
