+++ hw07_file_copying/go.mod
@@ -4,6 +4,5 @@ go 1.16
 
 require (
 	github.com/approvals/go-approval-tests v0.0.0-20220530063708-32d5677069bd
-	github.com/cheggaaa/pb/v3 v3.0.8
 	github.com/stretchr/testify v1.8.0
 )
