+++ hw07_file_copying/mycopy/utils_test.go
@@ -1,4 +1,4 @@
-package copy_test
+package mycopy_test
 
 import (
 	"os"
