+++ hw07_file_copying/progressbar/reader.go
@@ -1,14 +1,14 @@
-package progress_bar
+package progressbar
 
 import "io"
 
-// Reader is a wrapper around the std reader with a Bar embedded
+// Reader is a wrapper around the std reader with a Bar embedded.
 type Reader struct {
 	io.Reader
 	bar *Bar
 }
 
-// Read spies on the Reader and adds to the Bar
+// Read spies on the Reader and adds to the Bar.
 func (r *Reader) Read(p []byte) (int, error) {
 	n, err := r.Reader.Read(p)
 	if err != nil {
